Convert tests to xunit
