+++ AspNetCoreInjection.TypedFactories.Test/xunit.runner.json
@@ -1,0 +1,3 @@
+﻿{
+  "shadowCopy": false
+}
